refactor: Extract components from trading_cycle.py (#439)
Extract four components to improve maintainability and reduce complexity:

- LocalStateManager: JSON state persistence (140 lines)
- BrokerStateCache: Cached broker state with TTL (278 lines)
- StateReconciler: Reconciliation logic and alerts (475 lines)
- ReportGenerator: Human-readable reports (251 lines)

trading_cycle.py reduced from 1248 to 510 lines (59% reduction).
Maintains backward compatibility via property accessors and method wrappers.
